+++ CHANGELOG.md
@@ -23,7 +23,9 @@ Types of changes:
 - `test_load_hf_components:test_get_model_hf()` to assert model type
 - `coverage` for reporting of functions covered with `pytest`
 - `tests/conftest.py` for pytest fixtures
-- Loading of local models in `load_hf_components:get_model_hf()`
+- `load_hf_components.py`
+  - Loading of local models in `get_model_hf()`
+  - Local saving and loading of HF Metric Builder Scripts in `get_metrics_to_load_objects_hf()`
 
 ### Changed